グループ作成後にメンバー追加実装

+++ src/components/addGroup/_friendList/friendListItem.tsx
@@ -1,10 +1,10 @@
 // libs
-import React, { useState } from 'react';
+import React from 'react';
 import { View, StyleSheet, TouchableHighlight, Image, Text } from 'react-native';
 
 
 // layouts
-import { CONTENT_WIDTH, MAIN_WHITE_COLOR, PROFILE_IMAGE_SIZE, STANDARD_FONT, ICON_SIZE,PROFILE_IMAGE_BORDER_RADIUS } from '../../../constants/layout'
+import { CONTENT_WIDTH, MAIN_WHITE_COLOR, PROFILE_IMAGE_SIZE, STANDARD_FONT, ICON_SIZE, PROFILE_IMAGE_BORDER_RADIUS, MAIN_GRAY_COLOR } from '../../../constants/layout'
 
 
 export function FriendListItem({
@@ -12,6 +12,7 @@ export function FriendListItem({
 	addFriendList,
 	deleteFriendList,
 	selectedFriendList,
+	groupMemberUserId,
 }) {
 	let clicked = false
 	// 該当友達が選択されているかどうかの判定
@@ -23,14 +24,13 @@ export function FriendListItem({
 		<TouchableHighlight
 			onPress={() => {
 				// clickedがtrueの場合(この時点ではtrueに変わっていないので、falseで判定)
-				if (!clicked) {
+				if (!clicked && !groupMemberUserId.includes(list.friend_use_id)) {
 					addFriendList(list.key, list.type)
 				}
 				// clickedがfalseの場合(この時点ではfalseに変わっていないので、trueで判定)
-				if (clicked) {
+				if (clicked && !groupMemberUserId.includes(list.friend_use_id)) {
 					deleteFriendList(list.key, list.type)
 				}
-
 			}}
 			style={styles.rowFrontStyle}
 			underlayColor={'#feffff'}
@@ -38,14 +38,14 @@ export function FriendListItem({
 			<View style={styles.listBoxStyle}>
 				<View style={styles.listWrapperStyle}>
 					<View style={styles.listItemContainerStyle}>
-						<Image source={list.friend_profile_image} style={styles.profileImageStyle} />
-						<Text style={styles.listItemNameStyle}>{list.friend_nickname}</Text>
+						<Image source={list.friend_profile_image} style={groupMemberUserId.includes(list.friend_use_id) ? styles.selectedProfileImageStyle : styles.profileImageStyle} />
+						<Text style={groupMemberUserId.includes(list.friend_use_id) ? styles.selectedListItemNameStyle : styles.listItemNameStyle}>{list.friend_nickname}</Text>
 					</View>
 				</View>
 				<View style={styles.circleContainerStyle}>
 					{clicked ? (
 						<Image source={require("../../../../assets/images/checked-circle.png")} style={styles.circleStyle} />
-					) :
+					) : groupMemberUserId.includes(list.friend_use_id) ? null :
 						(
 							<Image source={require("../../../../assets/images/gray_circle.png")} style={styles.circleStyle} />
 						)}
@@ -61,6 +61,12 @@ export const styles = StyleSheet.create({
 		backgroundColor: MAIN_WHITE_COLOR,
 		justifyContent: 'center',
 		height: 50,
+	},
+	selectedProfileImageStyle: {
+		width: PROFILE_IMAGE_SIZE,
+		height: PROFILE_IMAGE_SIZE,
+		borderRadius: PROFILE_IMAGE_BORDER_RADIUS,
+		opacity: 0.5,
 	},
 	profileImageStyle: {
 		width: PROFILE_IMAGE_SIZE,
@@ -82,6 +88,11 @@ export const styles = StyleSheet.create({
 		display: "flex",
 		flexDirection: "row",
 		alignItems: "center",
+	},
+	selectedListItemNameStyle: {
+		fontFamily: STANDARD_FONT,
+		marginLeft: 12,
+		color: MAIN_GRAY_COLOR
 	},
 	listItemNameStyle: {
 		fontFamily: STANDARD_FONT,
